requirements.txt
python-dotenv
crewai
pandas==1.5.1
langchain-groq==0.0.1
pytrends==4.9.2 #unofficial api for google trends
beautifulsoup4==4.12.3#library to extract data from xml and html
groq==0.5.0
crewai_tools==0.2.5
pydantic==2.7.1
tweepy==4.14.0#to access twitter api
langchain
requests
groq